fix(electron): better support for custom schemas
