+++ tests/electron/assets/imported.css
@@ -1,0 +1,3 @@
+button {
+  font-weight: bold;
+}
