+++ tests/electron/assets/index.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <link rel="stylesheet" href="vscode-file://style.css">
+    <style>
+      @import url("vscode-file://imported.css");
+      button {
+        background: url(background.png);
+      }
+    </style>
+  </head>
+  <body>
+    <button>Click me</button>
+  </body>
+</html>
