+++ tests/electron/assets/style.css
@@ -1,0 +1,3 @@
+button {
+  color: red;
+}
