+++ tests/electron/electron-app.js
@@ -1,3 +1,11 @@
-const { app } = require('electron');
+const { app, protocol } = require('electron');
+const path = require('path');
 
 app.on('window-all-closed', e => e.preventDefault());
+
+app.whenReady().then(() => {
+  protocol.registerFileProtocol('vscode-file', (request, callback) => {
+    const url = request.url.substring('vscode-file'.length + 3);
+    callback({ path: path.join(__dirname, 'assets', url) });
+  });
+});
