+++ tests/electron/electronTest.ts
@@ -16,8 +16,9 @@
 
 import { baseTest } from '../config/baseTest';
 import * as path from 'path';
-import { ElectronApplication, Page } from 'playwright-core';
+import { ElectronApplication, Page } from '@playwright/test';
 import { PageTestFixtures, PageWorkerFixtures } from '../page/pageTestApi';
+import { traceViewerFixtures, TraceViewerFixtures } from '../config/traceViewerFixtures';
 export { expect } from '@playwright/test';
 
 type ElectronTestFixtures = PageTestFixtures & {
@@ -27,7 +28,7 @@ type ElectronTestFixtures = PageTestFixtures & {
 
 const electronVersion = require('electron/package.json').version;
 
-export const electronTest = baseTest.extend<ElectronTestFixtures, PageWorkerFixtures>({
+export const electronTest = baseTest.extend<TraceViewerFixtures>(traceViewerFixtures).extend<ElectronTestFixtures, PageWorkerFixtures>({
   browserVersion: [electronVersion, { scope: 'worker' }],
   browserMajorVersion: [Number(electronVersion.split('.')[0]), { scope: 'worker' }],
   isAndroid: [false, { scope: 'worker' }],
@@ -70,5 +71,9 @@ export const electronTest = baseTest.extend<ElectronTestFixtures, PageWorkerFixt
 
   page: async ({ newWindow }, run) => {
     await run(await newWindow());
+  },
+
+  context: async ({ electronApp }, run) => {
+    await run(electronApp.context());
   },
 });
